moving class instance to models
